Add vulkano-win crate and move out examples
