+++ examples/src/bin/image_fs.glsl
