+++ examples/src/bin/image_vs.glsl
