+++ examples/src/bin/teapot_fs.glsl
