+++ examples/src/bin/teapot_vs.glsl
