+++ examples/src/bin/triangle_fs.glsl
