+++ examples/src/bin/triangle_vs.glsl
